Refactor : 적 히트 혹은 사망시 전후좌우 애니메이션 분리

+++ Assets/02.Scripts/03.Enemy/Sub/Projectile/EnemyArrow.cs
@@ -4,7 +4,7 @@ public class EnemyArrow : EnemyProjectile
 {
     [SerializeField] HitBox _hitbox;
 
-    public override void Init(int damage)
+    public override void Init(float damage)
     {
         base.Init(damage);
         _hitbox?.Activate();
